L3: L3.6: dev: CreateSalesRecord

+++ L3/L3.6/sales-tracker/sales-tracker_main-server/internal/repository/rpPostgres/rpPostgresSales.go
@@ -9,7 +9,7 @@ import (
 
 // CreateSalesRecord creates a new sales record in the database
 func (rp *RpPostgres) CreateSalesRecord(ctx context.Context, data model.Data) (int, error) {
-	amountDB := data.Amount * 100
+	amountDB := int(data.Amount * 100)
 
 	query := `
